messagesender/requirements.txt
psycopg2-binary
sqlalchemy-json
sqlalchemy-timescaledb
jinja2
requests
